TOTAL DESIGN REVAMP! more in desc.
Changed home page and all navigation. 
This version is strictly only finished only for desktop, it is very much not responsive. 

There are now different product pages featuring the new graphics for, fruit, dark and ruby chocolates. 

Individual product page has been re-done with a slightly different navigation and flexbox tweaks. Will probably need breakpoints to prevent it from breaking and re-design the css elements. 

The new hamburger icon does need to be fixed graphically as well, I can also get on that on Wednesday.

'DivineHome.html' is the new home page and featured parallax background scrolling with our new graphic design and art direction. 

The nav-bar has been redone to now make a hamburger icon (which still needs to be re-designed, will probably do on Wednesday) 

We now have the 'back-to-top' button on all pages, since all of the new javascript is in one file we need it in all pages for it to work, otherwise there will be an error. 

TLDR; A lot of stuff added. :) 

Things to do / fix: 
MAJOR : 
Mobile responsiveness (just use breakpoints) 
+/- on items on their individual pages and in the cart. 
Minor: 
Shopping cart icon 
Hamburger icon and the menu 

Things to talk about: The text on the banners, should we take them out from the images and just use text? (We would have to import more text)